feat: Dropping a remote image will now upload
