+++ shared/editor/nodes/SimpleImage.tsx
@@ -11,7 +11,7 @@ import insertFiles, { Options } from "../commands/insertFiles";
 import { default as ImageComponent } from "../components/Image";
 import { MarkdownSerializerState } from "../lib/markdown/serializer";
 import uploadPlaceholderPlugin from "../lib/uploadPlaceholder";
-import uploadPlugin from "../lib/uploadPlugin";
+import { UploadPlugin } from "../plugins/UploadPlugin";
 import { ComponentProps } from "../types";
 import Node from "./Node";
 
@@ -238,6 +238,6 @@ export default class SimpleImage extends Node {
   }
 
   get plugins() {
-    return [uploadPlaceholderPlugin, uploadPlugin(this.options)];
+    return [uploadPlaceholderPlugin, new UploadPlugin(this.options)];
   }
 }
